+++ Mask-RCNN/paddle/rcnn/run.sh
@@ -34,7 +34,12 @@ train(){
    --MASK_ON=True > ${log_file} 2>&1 &
   train_pid=$!
   sleep 600
-  kill -9 $train_pid
+
+  line=`ps aux | grep ${train_pid} |grep -v "grep"|wc -l`
+  if [ $line -gt 0 ]; 
+  then
+     kill -9 $train_pid
+  fi
 }
 
 infer(){
